Retrieve reply message from discord client
According to docs, the client always echoes back a package that we can
use for debugging, so fetch it. Since I/O is synchronous, we have to
rely on the discord client to behave. In order not to potentially block
ST's UI from commands, wrap them in set_timeout_async.

Closes #27.